Fixed loading of .data in linker scripts.
This fixes a problem where the linker included some padding
bytes between the end of the .text section (_etext) and the
start of the .data section.

The C runtime copies from _etext, so all static initialised data
was corrupted.  This change forces the .data section to be written
at _etext to avoid this problem.