examples: Remove duplicate example linker scripts.
Only add one linker script per eval board, and have the Makefiles use that.

This avoids duplication, and various copy-paste errors and left-overs
which we already have in the tree.

Some of the linker scripts are not really correct, yet. Both, the comments
which say which board, chip, and RAM/ROM sizes are used, as well as the
actual RAM/ROM size definitions are sometimes copy-pasted from other boards
and are thus incorrect. This needs to be fixed.