[Doxygen] Strip out architectures from Makefile dependencies
This commit shrinks the Makefile of the doxygen and makes easy error-prone addition of the new supported platforms.
The user should modify only one line at start of file to add new supported family to the build process.

The calling format is preserved and is compatible with old version of the makefile.