[stm32f4] Removed the rcc_apbx_clock variables in favor of rcc_clock_scale_used.
We are now storing just the struct we used to setup the clock. It
contains the resulting clock values anyways and additionally we can
retrieve all other settings too. At the end it is more memory efficient
anyways as most of the values stay in flash and we only use one single
pointer to remember what we used for the future.