Super EZ Clock Setup
The goal of this commit was to introduce the function rcc_clock_setup() which
takes exactly two parameters, desired clock rate and the hse clock rate (if
HSE is available) zero otherwise. It figures out what to do and which
clock to use based on what it knows about the chip.