stm32/quadspi: Add byte addressable define
The data register can be accessed as a byte, a word, or a long.
It is most easily accessed as a byte when using byte oriented
SPI devices (like Flash chips and display chips). By having the
byte oriented version you can drain the fifo more effectively
than if you are trying to match the current number of bytes in
the FIFO versus the number you would get if your read the register
as word or a long.