makefile: make it easier to disable targets.
Sometimes you're working on a subset of targets, and want to do full
clean/builds, but don't need to clean and rebuild families you're not
working with.  This is a pure syntactic change that makes it easier to
simply comment out lines for targets, instead of having to mangle line
continuations.