STM32L: USART: Fix baudrate calculation for LPUART1
LPUART1 needs BRR to be 256 times the normal value.

To calculate this correctly with rounding, the best answer
is to use 64 bit values: A clock of say 20 MHz times 256 is
already 5e9 MHz, which doesn't fit in 32 bits.

There are probably some tricks using GCD, fixed point math
or something, but really, a little 64 bit math wont hurt.
This is setup code, which usually is not performance
critical.