Added support for mandatory vital product data (VPD) pages. Added basic support for optional Unit Serial Number VPD.
According to SPC-3 two vital product data pages are mandatory:
The "Supported VPD Pages" page and the "Device Identification" page.
This change implements these pages using the minimum of information
required by the specification.

The optional VPD page "Unit Serial Number" is supported since
it is queried by the Windows host operating system. Again, a minimum
implementation is provided, the serial number is reported as a single
space character, signaling that the serial number is not available.