genlink: Provide LIBNAME even if library isn't built yet
Currently genlink-config.mk tries to look for the device library by first
looking for the family and then subfamily suffix. This has two problems:

1) If the library wasn't built yet we get a warning and LIBNAME isn't
set. This makes building libopencm3 automatically as part of another
project's makefile uneccesarily difficult as LIBNAME has to be hardcoded as a
workaround.

2) In the special case of the lpc43xx_m0 "device" the correct LIBNAME is
actually the one with subfamily suffix, but currently we prefere the family
one.

Both of these issues are fixed by makeing the canonical LIBNAME part of the
genlink's database. When asking genlink for a LIBNAME we default to the
family suffix so as to not require annotating evey device manually but
allow overriding this by specifying a LIBNAME as a definition in
devices.data.

Currently lpc43xx_m0 is the only odd one out that requires this though.