add library target for f3_nofpu
A few early batches of the gd32f303 had no FPU. To use these devices set
DEVICE to gd32f303cc_nofpu (for example). The presence of the FPU can be
detected at runtime by attempting to set bits 0xf00000 in SCB_CPACR and
checking whether the value is retained.