gitignore: ensure some non-generated are included
A normal git clone doesn't catch these, as it _knows_ that the files are
in the git repo and should be tracked.  However, downloading a tarball,
and adding it to a new git repo shows up the problem as important files
are simply not added, and would be dropped by git clean.

Reported-by: https://github.com/libopencm3/libopencm3/issues/1153