STM32H7 ADC Definitions and Device IDs
Addition of all register definitions for the H7 ADC and modifications to
existing ADC code where needed to support the fact that the register set
is ever so slightly different as needed to support 16bit and oversampled
samples.

Code across the ADC land was also changed to support const correctness.

And because the H7 ADC differs depending on the silicon revision, the
DBGMCU header has been massaged so that device ID and revision can be
extracted using known identifiers.